// grafana-plugin/src/api/generated/pluginBackend.ts
export interface Response {
  code: number;
  message?: string;
  data?: unknown;
}